Initial (late) push
This looks to be an ongoing assignment for CS_4330, so I am deciding to
try using source control for it. I am actually in the middle of
altering my implementation to closer match what the professor has in
his solution for the first checkpoint.
Currently, I can:
-Dynamically change the size of the board by dragging the border
-Choose between two color schemes
-Choose different grid layouts (8x8, 10x10, etc)

Right now I am changing the fxml portion to better prepare for later
adding the game pieces. I also need to get the board centered in the
scene whenever the window is not square, as opposed to being in the
upper left corner all the time.